REFACTOR : 출퇴근 기록, 지출 타입 경로를 domain으로 변경
