+++ src/main/java/com/piaar_store_manager/server/domain/account_book/proj/AccountBookJoinProj.java
@@ -1,8 +1,8 @@
 package com.piaar_store_manager.server.domain.account_book.proj;
 
 import com.piaar_store_manager.server.domain.account_book.entity.AccountBookEntity;
+import com.piaar_store_manager.server.domain.expenditure_type.entity.ExpenditureTypeEntity;
 import com.piaar_store_manager.server.domain.user.entity.UserEntity;
-import com.piaar_store_manager.server.model.expenditure_type.entity.ExpenditureTypeEntity;
 
 public interface AccountBookJoinProj {
     AccountBookEntity getAccountBook();
