+++ src/main/java/com/piaar_store_manager/server/domain/commute_record/repository/CommuteRecordRepository.java
@@ -1,6 +1,5 @@
-package com.piaar_store_manager.server.model.commute_record.repository;
+package com.piaar_store_manager.server.domain.commute_record.repository;
 
-import com.piaar_store_manager.server.model.commute_record.entity.CommuteRecordEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
@@ -9,6 +8,8 @@ import java.util.Date;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
+
+import com.piaar_store_manager.server.domain.commute_record.entity.CommuteRecordEntity;
 
 @Repository
 public interface CommuteRecordRepository extends JpaRepository<CommuteRecordEntity, Integer> {
